--- src/main/java/roomescape/common/utils/SecretKeyGenerator.java
@@ -1,28 +1,0 @@
-package roomescape.common.utils;
-
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.util.Base64;
-
-public class SecretKeyGenerator {
-    private static final int KEY_LENGTH = 104;
-
-    private static String generateKey() {
-        SecureRandom secureRandom;
-        try {
-            secureRandom = SecureRandom.getInstanceStrong();
-        }
-        catch (NoSuchAlgorithmException e) {
-            throw new RuntimeException(e);
-        }
-
-        byte[] keyBytes = new byte[KEY_LENGTH];
-        secureRandom.nextBytes(keyBytes);
-
-        return Base64.getEncoder().encodeToString(keyBytes);
-    }
-
-    public static void main(String[] args) {
-        System.out.println("발급된 비밀 키 : " + generateKey());
-    }
-}
